+++ src/app/scripts/validateAlbum.js
@@ -1,38 +1,15 @@
 import Ajv from 'ajv'
-import isIpfs from 'is-ipfs'
-import schemaObj from '../data/albums/schema'
+import albums from '../data/albums'
 
 const validateAlbum = (albumCandidate) => {
   const validator = new Ajv({
     allErrors: true
   })
-  let valid
-  const validatorErrors = []
-  valid = validator.validate(schemaObj, albumCandidate)
-  if (!valid) {
-    validatorErrors.push(...validator.errors)
-  }
-  const { cover, tracks } = albumCandidate
-  const coverIsMultihash = isIpfs.multihash(cover)
-  if (!coverIsMultihash) {
-    valid = false
-    validatorErrors.push({
-      dataPath: '.cover',
-      message: 'Must be a multihash value.'
-    })
-  }
-  tracks.forEach(({ hash }, index) => {
-    const isMultihash = isIpfs.multihash(hash)
-    if (!isMultihash) {
-      valid = false
-      validatorErrors.push({
-        dataPath: `tracks[${index}].hash`,
-        message: 'Must be a multihash value.'
-      })
-    }
-  })
+  const valid = validator.validate(albums.instanceSchema, albumCandidate)
+  console.log(valid)
+  console.log(validator.errors)
   return {
-    valid, validatorErrors
+    valid, errors: validator.errors
   }
 }