+++ src/app/utils/cidToBuffer.js
@@ -1,0 +1,5 @@
+const multibase = require('multibase')
+
+export default (cid) => {
+  return multibase.decode(cid)
+}