+++ src/app/utils/getCidString.js
@@ -2,6 +2,6 @@ const CID = require('cids')
 
 export default (incom) => {
   if (typeof incom === 'string') return incom
-  const cidObj = new CID(1, 'dag-cbor', incom)
+  const cidObj = new CID(incom)
   return cidObj.toBaseEncodedString()
 }