+++ src/app/view/Playlist/Playlist.js
@@ -1,4 +1,3 @@
-import React from 'react'
 import bind from '../../utils/recallReact'
 import playlistState from '../../state/playlist'
 import playerState from '../../state/player'