+++ CHANGELOG.md
@@ -2,6 +2,32 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.4.0"></a>
+# [0.4.0](https://github.com/pathephone/pathephone-desktop/compare/v0.2.3...v0.4.0) (2017-12-18)
+
+
+### Bug Fixes
+
+* **mac:** main window appears on Mac OS ([297fdb1](https://github.com/pathephone/pathephone-desktop/commit/297fdb1)), closes [#50](https://github.com/pathephone/pathephone-desktop/issues/50)
+
+
+### Features
+
+* **forms:** tracks autofill from id3 ([4a50555](https://github.com/pathephone/pathephone-desktop/commit/4a50555)), closes [#53](https://github.com/pathephone/pathephone-desktop/issues/53)
+* **playlist:** download full playlist in background ([aa99d1e](https://github.com/pathephone/pathephone-desktop/commit/aa99d1e))
+
+
+
+<a name="0.3.0"></a>
+# [0.3.0](https://github.com/pathephone/pathephone-desktop/compare/v0.2.3...v0.3.0) (2017-12-17)
+
+
+### Features
+
+* **forms:** tracks autofill from id3 ([4a50555](https://github.com/pathephone/pathephone-desktop/commit/4a50555)), closes [#53](https://github.com/pathephone/pathephone-desktop/issues/53)
+
+
+
 <a name="0.2.3"></a>
 ## [0.2.3](https://github.com/pathephone/pathephone-desktop/compare/v0.2.2...v0.2.3) (2017-12-03)